1. LearningManager.java 2. Extract Class 3. 해당 파일 내부에 클래스가 총 3개가 있어서 OCP를 위반하므로 이를 해결하기 위해 분리.
